add test_grpc_server
